// src/types/task-status/enums.ts
export enum TaskStatus {
  _SELECT_ = '_SELECT_',
  REGISTRATION_OK_NO_EMAIL = 'REGISTRATION_OK_NO_EMAIL',
  REGISTRATION_OK_WITH_EMAIL = 'REGISTRATION_OK_WITH_EMAIL',
  REGISTRATION_PENDING = 'REGISTRATION_PENDING',
  REGISTRATION_PENDING_VALIDATED = 'REGISTRATION_PENDING_VALIDATED',
  REGISTRATION_PENDING_DENIED = 'REGISTRATION_PENDING_DENIED',
  REGISTRATION_DENIED = 'REGISTRATION_DENIED',
  FORUM_NOT_WORKING = 'FORUM_NOT_WORKING',
  NO_FORUM = 'NO_FORUM',
  DIRECT_MESSAGE = 'DIRECT_MESSAGE',
  PROBLEM = 'PROBLEM',
  BANNED = 'BANNED',
  OTHER_PROBLEM = 'OTHER_PROBLEM'
}

export enum LinkVerificationStatus {
  _SELECT_ = '_SELECT_',
  PRESENT = 'PRESENT',
  NOT_PRESENT = 'NOT_PRESENT',
  BANNED = 'BANNED',
  OTHER_PROBLEM = 'OTHER_PROBLEM'
}

export enum IndexationStatus {
  _SELECT_ = '_SELECT_',
  NOT_INDEXED = 'NOT_INDEXED',
  IN_PROGRESS = 'IN_PROGRESS',
  INDEXED = 'INDEXED'
}

export enum TaskProcessStatus {
  INSCRIPTION = 'INSCRIPTION',
  INSCRIPTION_EN_ATTENTE = 'INSCRIPTION_EN_ATTENTE',
  POST_MESSAGE = 'POST_MESSAGE',
  AJOUT_LIEN = 'AJOUT_LIEN',
  VERIFICATION_LIEN = 'VERIFICATION_LIEN',
  INDEXATION = 'INDEXATION',
  TACHE_TERMINEE = 'TACHE_TERMINEE',
  TACHE_PROBLEME = 'TACHE_PROBLEME'
}